Add clustering, close #18
First of all, I had to add two new scripts:
 - MarkerCluster: https://github.com/Leaflet/Leaflet.markercluster
 - SubGroup: https://github.com/ghybs/Leaflet.FeatureGroup.SubGroup
Since we are still on Leaflet 0.7 I chose the corresponding releases
in both cases, rather than the latest releases.

In the link element I did not include the integrity attribute
as I don't know how to find out the hashes.

The implementation is rather straightforward, in `main.js` not much
had to be added.