Add events again
Now reads the manually edited list of events from a forum topic instead of an event plugin.
Filter out past events in case the data source wasn't updated and contains old events.
Sort events icons by date to have the newest one on top in case icons overlap because the events are very close or in the same location.